requirements.txt
motor>=3.1.0
python-telegram-bot==20.3
google-generativeai>=0.3.0
python-dotenv>=0.19.0
aiohttp>=3.8.0